Wait for cilium-ready before writing cni config.
The netd-init container currently writes the cni and this signals node-readiness
to the kubelet.
However, the current check is only for the presence of the cilium-cni file, which
is not a guarantee of cilium pod's readiness.
From the code, it seems that if the healthz returns error-code 200, then this
indicates that all cilium subsystems are ok, which is a better signal for
node-readiness.

This was tested by uploading a private a image and testing it in a sandbox
cluster by resizing the number of nodes in a cluster and verifying the logs.