Read CILIUM_CHAINING_MODE env var and use it for cilium-cni config
In the form of {"type": "cilium-cni", "chaining-mode": "generic-veth"}.

This was added in cilium/cilium#24956 as a required field when chaining,
since Cilium 1.14. We didn't run into issues without it because we carry
internal patches around the conditions, and happen to bypass the check.
However this field is required when using vanilla Cilium OSS, and our
future migration may want it too. Create a toggle for setting it.